requirements.txt
Flask
numpy
pandas
psycopg2-binary
python-dotenv
requests